1. add test cases for parsing command and columns, 2. refactor test suites
