+++ sql/hive/src/test/scala/org/apache/spark/sql/hive/StatisticsSuite.scala
@@ -18,42 +18,19 @@
 package org.apache.spark.sql.hive
 
 import java.io.{File, PrintWriter}
-import java.sql.{Date, Timestamp}
 
 import scala.reflect.ClassTag
 
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{AnalysisException, QueryTest, Row}
+import org.apache.spark.sql.{AnalysisException, Row}
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.plans.logical.{BasicColStats, Statistics}
 import org.apache.spark.sql.execution.command.{AnalyzeTableCommand, DDLUtils}
-import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.execution.joins._
-import org.apache.spark.sql.hive.test.TestHiveSingleton
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.test.SQLTestUtils
-import org.apache.spark.sql.types.{StructType, _}
+import org.apache.spark.sql.types.StructType
 
-class StatisticsSuite extends QueryTest with TestHiveSingleton with SQLTestUtils {
+class StatisticsSuite extends StatisticsTest {
 
   test("parse analyze commands") {
-    def assertAnalyzeCommand(analyzeCommand: String, c: Class[_]) {
-      val parsed = spark.sessionState.sqlParser.parsePlan(analyzeCommand)
-      val operators = parsed.collect {
-        case a: AnalyzeTableCommand => a
-        case o => o
-      }
-
-      assert(operators.size === 1)
-      if (operators(0).getClass() != c) {
-        fail(
-          s"""$analyzeCommand expected command: $c, but got ${operators(0)}
-             |parsed command:
-             |$parsed
-           """.stripMargin)
-      }
-    }
-
     assertAnalyzeCommand(
       "ANALYZE TABLE Table1 COMPUTE STATISTICS",
       classOf[AnalyzeTableCommand])
@@ -171,39 +148,6 @@ class StatisticsSuite extends QueryTest with TestHiveSingleton with SQLTestUtils
     }
     spark.sessionState.catalog.dropTable(
       TableIdentifier("tempTable"), ignoreIfNotExists = true, purge = false)
-  }
-
-  private def checkTableStats(
-      stats: Option[Statistics],
-      hasSizeInBytes: Boolean,
-      expectedRowCounts: Option[Int]): Unit = {
-    if (hasSizeInBytes || expectedRowCounts.nonEmpty) {
-      assert(stats.isDefined)
-      assert(stats.get.sizeInBytes > 0)
-      assert(stats.get.rowCount === expectedRowCounts)
-    } else {
-      assert(stats.isEmpty)
-    }
-  }
-
-  private def checkTableStats(
-      tableName: String,
-      isDataSourceTable: Boolean,
-      hasSizeInBytes: Boolean,
-      expectedRowCounts: Option[Int]): Option[Statistics] = {
-    val df = sql(s"SELECT * FROM $tableName")
-    val stats = df.queryExecution.analyzed.collect {
-      case rel: MetastoreRelation =>
-        checkTableStats(rel.catalogTable.stats, hasSizeInBytes, expectedRowCounts)
-        assert(!isDataSourceTable, "Expected a Hive serde table, but got a data source table")
-        rel.catalogTable.stats
-      case rel: LogicalRelation =>
-        checkTableStats(rel.catalogTable.get.stats, hasSizeInBytes, expectedRowCounts)
-        assert(isDataSourceTable, "Expected a data source table, but got a Hive serde table")
-        rel.catalogTable.get.stats
-    }
-    assert(stats.size == 1)
-    stats.head
   }
 
   test("test table-level statistics for hive tables created in HiveExternalCatalog") {
@@ -337,229 +281,6 @@ class StatisticsSuite extends QueryTest with TestHiveSingleton with SQLTestUtils
         isDataSourceTable = true,
         hasSizeInBytes = true,
         expectedRowCounts = Some(10))
-    }
-  }
-
-  private def checkColStats(
-      rowRDD: RDD[Row],
-      schema: StructType,
-      expectedColStatsSeq: Seq[(String, BasicColStats)]): Unit = {
-    val table = "tbl"
-    withTable(table) {
-      var df = spark.createDataFrame(rowRDD, schema)
-      df.write.format("json").saveAsTable(table)
-      val columns = expectedColStatsSeq.map(_._1).mkString(", ")
-      sql(s"ANALYZE TABLE $table COMPUTE STATISTICS FOR COLUMNS $columns")
-      df = sql(s"SELECT * FROM $table")
-      val stats = df.queryExecution.analyzed.collect {
-        case rel: LogicalRelation =>
-          expectedColStatsSeq.foreach { expected =>
-            assert(rel.catalogTable.get.stats.get.basicColStats.contains(expected._1))
-            checkColStats(colStats = rel.catalogTable.get.stats.get.basicColStats(expected._1),
-              expectedColStats = expected._2)
-          }
-      }
-      assert(stats.size == 1)
-    }
-  }
-
-  private def checkColStats(colStats: BasicColStats, expectedColStats: BasicColStats): Unit = {
-    assert(colStats.dataType == expectedColStats.dataType)
-    assert(colStats.numNulls == expectedColStats.numNulls)
-    colStats.dataType match {
-      case ByteType | ShortType | IntegerType | LongType =>
-        assert(colStats.max.map(_.toString.toLong) == expectedColStats.max.map(_.toString.toLong))
-        assert(colStats.min.map(_.toString.toLong) == expectedColStats.min.map(_.toString.toLong))
-      case FloatType | DoubleType =>
-        assert(colStats.max.map(_.toString.toDouble) == expectedColStats.max
-          .map(_.toString.toDouble))
-        assert(colStats.min.map(_.toString.toDouble) == expectedColStats.min
-          .map(_.toString.toDouble))
-      case DecimalType.SYSTEM_DEFAULT =>
-        assert(colStats.max.map(i => Decimal(i.toString)) == expectedColStats.max
-          .map(i => Decimal(i.toString)))
-        assert(colStats.min.map(i => Decimal(i.toString)) == expectedColStats.min
-          .map(i => Decimal(i.toString)))
-      case DateType | TimestampType =>
-        if (expectedColStats.max.isDefined) {
-          // just check the difference to exclude the influence of timezones
-          assert(colStats.max.get.toString.toLong - colStats.min.get.toString.toLong ==
-            expectedColStats.max.get.toString.toLong - expectedColStats.min.get.toString.toLong)
-        } else {
-          assert(colStats.max.isEmpty && colStats.min.isEmpty)
-        }
-      case _ => // only numeric types, date type and timestamp type have max and min stats
-    }
-    colStats.dataType match {
-      case BinaryType => assert(colStats.ndv.isEmpty)
-      case BooleanType => assert(colStats.ndv.contains(2))
-      case _ =>
-        // ndv is an approximate value, so we just make sure we have the value
-        assert(colStats.ndv.get >= 0)
-    }
-    assert(colStats.avgColLen == expectedColStats.avgColLen)
-    assert(colStats.maxColLen == expectedColStats.maxColLen)
-    assert(colStats.numTrues == expectedColStats.numTrues)
-    assert(colStats.numFalses == expectedColStats.numFalses)
-  }
-
-  test("basic statistics for integral type columns") {
-    val rdd = sparkContext.parallelize(Seq("1", null, "2", "3", null)).map { i =>
-      if (i != null) Row(i.toByte, i.toShort, i.toInt, i.toLong) else Row(i, i, i, i)
-    }
-    val schema = StructType(
-      StructField(name = "c1", dataType = ByteType, nullable = true) ::
-      StructField(name = "c2", dataType = ShortType, nullable = true) ::
-      StructField(name = "c3", dataType = IntegerType, nullable = true) ::
-      StructField(name = "c4", dataType = LongType, nullable = true) :: Nil)
-    val expectedBasicStats = BasicColStats(
-      dataType = ByteType, numNulls = 2, max = Some(3), min = Some(1), ndv = Some(3))
-    val statsSeq = Seq(
-      ("c1", expectedBasicStats),
-      ("c2", expectedBasicStats.copy(dataType = ShortType)),
-      ("c3", expectedBasicStats.copy(dataType = IntegerType)),
-      ("c4", expectedBasicStats.copy(dataType = LongType)))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for fractional type columns") {
-    val rdd = sparkContext.parallelize(Seq(null, "1.01", "2.02", "3.03")).map { i =>
-      if (i != null) Row(i.toFloat, i.toDouble, Decimal(i)) else Row(i, i, i)
-    }
-    val schema = StructType(
-      StructField(name = "c1", dataType = FloatType, nullable = true) ::
-      StructField(name = "c2", dataType = DoubleType, nullable = true) ::
-      StructField(name = "c3", dataType = DecimalType.SYSTEM_DEFAULT, nullable = true) :: Nil)
-    val expectedBasicStats = BasicColStats(
-      dataType = FloatType, numNulls = 1, max = Some(3.03), min = Some(1.01), ndv = Some(3))
-    val statsSeq = Seq(
-      ("c1", expectedBasicStats),
-      ("c2", expectedBasicStats.copy(dataType = DoubleType)),
-      ("c3", expectedBasicStats.copy(dataType = DecimalType.SYSTEM_DEFAULT)))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for string column") {
-    val rdd = sparkContext.parallelize(Seq(null, "a", "bbbb", "cccc")).map(Row(_))
-    val schema = StructType(StructField(name = "c1", dataType = StringType, nullable = true) :: Nil)
-    val statsSeq = Seq(("c1", BasicColStats(dataType = StringType, numNulls = 1,
-      maxColLen = Some(4), avgColLen = Some(2.25), ndv = Some(3))))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for binary column") {
-    val rdd = sparkContext.parallelize(Seq(null, "a", "bbbb", "cccc")).map { i =>
-      if (i != null) Row(i.getBytes) else Row(i)
-    }
-    val schema = StructType(StructField(name = "c1", dataType = BinaryType, nullable = true) :: Nil)
-    val statsSeq = Seq(("c1", BasicColStats(dataType = BinaryType, numNulls = 1,
-      maxColLen = Some(4), avgColLen = Some(2.25))))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for boolean column") {
-    val rdd = sparkContext.parallelize(Seq(null, true, false, true)).map(Row(_))
-    val schema =
-      StructType(StructField(name = "c1", dataType = BooleanType, nullable = true) :: Nil)
-    val statsSeq = Seq(("c1", BasicColStats(dataType = BooleanType, numNulls = 1,
-      numTrues = Some(2), numFalses = Some(1))))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for date column") {
-    val rdd = sparkContext.parallelize(Seq(null, "1970-01-01", "1970-02-02")).map { i =>
-      if (i != null) Row(Date.valueOf(i)) else Row(i)
-    }
-    val schema =
-      StructType(StructField(name = "c1", dataType = DateType, nullable = true) :: Nil)
-    val statsSeq = Seq(("c1", BasicColStats(dataType = DateType, numNulls = 1,
-      max = Some(32), min = Some(0), ndv = Some(2))))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for timestamp column") {
-    val rdd = sparkContext.parallelize(Seq(null, "1970-01-01 00:00:00", "1970-01-01 00:00:05"))
-      .map(i => if (i != null) Row(Timestamp.valueOf(i)) else Row(i))
-    val schema =
-      StructType(StructField(name = "c1", dataType = TimestampType, nullable = true) :: Nil)
-    val statsSeq = Seq(("c1", BasicColStats(dataType = TimestampType, numNulls = 1,
-      max = Some(5000000), min = Some(0), ndv = Some(2))))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for null columns") {
-    val rdd = sparkContext.parallelize(Seq(Row(null, null)))
-    val schema = StructType(
-      StructField(name = "c1", dataType = LongType, nullable = true) ::
-      StructField(name = "c2", dataType = TimestampType, nullable = true) :: Nil)
-    val expectedBasicStats = BasicColStats(dataType = LongType, numNulls = 1,
-      max = None, min = None, ndv = Some(0))
-    val statsSeq = Seq(
-      ("c1", expectedBasicStats),
-      ("c2", expectedBasicStats.copy(dataType = TimestampType)))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("basic statistics for columns with different types") {
-    val rdd = sparkContext.parallelize(Seq(
-      Row(1, 1.01, "a", "a".getBytes, true, Date.valueOf("1970-01-01"),
-        Timestamp.valueOf("1970-01-01 00:00:00"), 5.toLong),
-      Row(2, 2.02, "bb", "bb".getBytes, false, Date.valueOf("1970-02-02"),
-        Timestamp.valueOf("1970-01-01 00:00:05"), 4.toLong)))
-    val schema = StructType(Seq(
-      StructField(name = "c1", dataType = IntegerType, nullable = false),
-      StructField(name = "c2", dataType = DoubleType, nullable = false),
-      StructField(name = "c3", dataType = StringType, nullable = false),
-      StructField(name = "c4", dataType = BinaryType, nullable = false),
-      StructField(name = "c5", dataType = BooleanType, nullable = false),
-      StructField(name = "c6", dataType = DateType, nullable = false),
-      StructField(name = "c7", dataType = TimestampType, nullable = false),
-      StructField(name = "c8", dataType = LongType, nullable = false)))
-    val statsSeq = Seq(
-      ("c1", BasicColStats(dataType = IntegerType, numNulls = 0, max = Some(2), min = Some(1),
-        ndv = Some(2))),
-      ("c2", BasicColStats(dataType = DoubleType, numNulls = 0, max = Some(2.02), min = Some(1.01),
-        ndv = Some(2))),
-      ("c3", BasicColStats(dataType = StringType, numNulls = 0, maxColLen = Some(2),
-        avgColLen = Some(1.5), ndv = Some(2))),
-      ("c4", BasicColStats(dataType = BinaryType, numNulls = 0, maxColLen = Some(2),
-        avgColLen = Some(1.5))),
-      ("c5", BasicColStats(dataType = BooleanType, numNulls = 0, numTrues = Some(1),
-        numFalses = Some(1), ndv = Some(2))),
-      ("c6", BasicColStats(dataType = DateType, numNulls = 0, max = Some(32), min = Some(0),
-        ndv = Some(2))),
-      ("c7", BasicColStats(dataType = TimestampType, numNulls = 0, max = Some(5000000),
-        min = Some(0), ndv = Some(2))),
-      ("c8", BasicColStats(dataType = LongType, numNulls = 0, max = Some(5), min = Some(4),
-        ndv = Some(2))))
-    checkColStats(rdd, schema, statsSeq)
-  }
-
-  test("update table-level stats while collecting column-level stats") {
-    val table = "tbl"
-    val tmpTable = "tmp"
-    withTable(table, tmpTable) {
-      val rdd = sparkContext.parallelize(Seq(Row(1)))
-      val df = spark.createDataFrame(rdd, StructType(Seq(
-        StructField(name = "c1", dataType = IntegerType, nullable = false))))
-      df.write.format("json").saveAsTable(tmpTable)
-
-      sql(s"CREATE TABLE $table (c1 int)")
-      sql(s"INSERT INTO $table SELECT * FROM $tmpTable")
-      sql(s"ANALYZE TABLE $table COMPUTE STATISTICS")
-      val fetchedStats1 = checkTableStats(tableName = table, isDataSourceTable = false,
-        hasSizeInBytes = true, expectedRowCounts = Some(1))
-
-      // update table between analyze table and analyze column commands
-      sql(s"INSERT INTO $table SELECT * FROM $tmpTable")
-      sql(s"ANALYZE TABLE $table COMPUTE STATISTICS FOR COLUMNS c1")
-      val fetchedStats2 = checkTableStats(tableName = table, isDataSourceTable = false,
-        hasSizeInBytes = true, expectedRowCounts = Some(2))
-
-      assert(fetchedStats2.get.sizeInBytes > fetchedStats1.get.sizeInBytes)
-      val basicColStats = fetchedStats2.get.basicColStats("c1")
-      checkColStats(colStats = basicColStats, expectedColStats = BasicColStats(
-        dataType = IntegerType, numNulls = 0, max = Some(1), min = Some(1), ndv = Some(1)))
     }
   }
